+++ src/models/mod.rs
@@ -7,6 +7,7 @@ use surrealdb::sql::Thing;
 pub mod access_token;
 pub mod auth_for;
 pub mod datetime;
+pub mod embeddings;
 pub mod file_metadata;
 pub mod refresh_token;
 pub mod session;
@@ -15,6 +16,7 @@ pub mod user_info;
 
 pub(crate) use access_token::*;
 pub(crate) use auth_for::*;
+pub(crate) use embeddings::*;
 pub(crate) use file_metadata::*;
 pub(crate) use refresh_token::*;
 pub(crate) use session::*;
